+++ app/classes/controllers/management_controller.py
@@ -164,9 +164,18 @@ class ManagementController:
         excluded_dirs: list = None,
         compress: bool = False,
         shutdown: bool = False,
+        before: str = "",
+        after: str = "",
     ):
         return self.management_helper.set_backup_config(
-            server_id, backup_path, max_backups, excluded_dirs, compress, shutdown
+            server_id,
+            backup_path,
+            max_backups,
+            excluded_dirs,
+            compress,
+            shutdown,
+            before,
+            after,
         )
 
     @staticmethod