+++ app/classes/shared/server.py
@@ -1024,7 +1024,17 @@ class ServerInstance:
             )
         time.sleep(3)
         conf = HelpersManagement.get_backup_config(self.server_id)
+        if conf["before"]:
+            if self.check_running():
+                logger.debug(
+                    "Found running server and send command option. Sending command"
+                )
+                self.send_command(conf["before"])
+
         if conf["shutdown"]:
+            if conf["before"]:
+                # pause to let people read message.
+                time.sleep(5)
             logger.info(
                 "Found shutdown preference. Delaying"
                 + "backup start. Shutting down server."
@@ -1105,6 +1115,14 @@ class ServerInstance:
                 self.run_threaded_server(HelperUsers.get_user_id_by_name("system"))
             time.sleep(3)
             self.last_backup_failed = False
+            if conf["after"]:
+                if self.check_running():
+                    logger.debug(
+                        "Found running server and send command option. Sending command"
+                    )
+                    self.send_command(conf["after"])
+            # pause to let people read message.
+            time.sleep(5)
         except:
             logger.exception(
                 f"Failed to create backup of server {self.name} (ID {self.server_id})"